Add model attributes
Inspected results from artist search to determine the minimal attributes to keep.
We want the `spotify_id` to store distinct versions locally using `updateOrCreate`.
The name is table stakes.
We filter for the largest image url to store and shrink it down to save time
Get the follower count
Store when the artist was last searched (and found)