Add repository and transformer
Copy .env to testing to isolate tests. Testing can blow away your local database if you're not careful.
Remove example tests.
Add a channel for logging as I had issues immediately when spiking.
Add repository that uses the Spotify service to search for a list of artist, filter by exact name (I need to lowercase this) and type to get the first. Ideally we'd sort by highest follower count to get the most popular choice.
Add transformer which converts the Spotify array into attributes for our model creation.
Fix controller test to be a todo sometime later. I'll likely test the moving pieces and may skip this entirely.
Add tests for the happy paths, mocking the SpotifyRequest as thankfully it's resolved instead of blindly set in a way we can't mock it out. We can reuse the same artist array and add the `artists.items` envelope since we're always expected to return this one artist.

I had mocking not quite right before and it took a minute to realize I was making calls to Spotify. I turned my wifi off to troubleshoot.
I tried mocking the Spotify fluid interface and it's not quite easy to understand method chaining. Finding which random project I've gotten the furthest with is a lesson in futility.
I tried using a PendingRequest object with parameters to return to the mocked call and it turns out that it makes an API call.
Typing this all out it makes sense to mock the actual request, not the pending one and it took a minute to get reacquainted with this library.