Add Track relationship
Add migration, factory, and seed.

Using the Artist seeder we can more easily use the magic method approach of adding artists with associated tracks.
Track seeder can create 20 individual tracks tied to new artists, but truncating here is a problem so we'll stick to the one seeder for now.

The `album_release_date` is the only known anomaly. It's a string, but I can't recall if I did a lot of normalizing to have it parse to a date time, i.e. this may change shortly.