Add indexes
Index the unique `spotify_id` columns that shouldn't change
Change images to nullable as some may not be present (or possibly the size isn't). I forgot this was true of other projects.

Looks like foreign keys do get indexes as `tracks_artist_id_foreign` was created but not unique which makes sense. This is a 1:M table, it'll have the same foreign key multiple times.